Update docker-compose to include build step
This build step sets up the Python application as a wheel package, ready to be used for acceptance testing and eventually as release candidate. This step required updates to test.sh script to download pip packages first before installing, so they could be used by both the test and build image creation steps. This makes both images the same, as they use the same downloaded packages saved in the cache.